kuring-69 [수정] notice 로컬, 서버 로직을 :data:notice로 옮김
