--- app/src/main/java/com/ku_stacks/ku_ring/data/api/response/DepartmentNoticeListResponse.kt
@@ -1,7 +1,0 @@
-package com.ku_stacks.ku_ring.data.api.response
-
-data class DepartmentNoticeListResponse(
-    val code: Int,
-    val message: String,
-    val data: List<DepartmentNoticeResponse>,
-)
