--- app/src/main/java/com/ku_stacks/ku_ring/data/api/response/DepartmentNoticeResponse.kt
@@ -1,10 +1,0 @@
-package com.ku_stacks.ku_ring.data.api.response
-
-data class DepartmentNoticeResponse(
-    val articleId: String?,
-    val postedDate: String?,
-    val subject: String?,
-    val url: String?,
-    val category: String?,
-    val important: Boolean,
-)
