--- app/src/main/java/com/ku_stacks/ku_ring/data/api/response/SearchNoticeDataResponse.kt
@@ -1,7 +1,0 @@
-package com.ku_stacks.ku_ring.data.api.response
-
-import com.google.gson.annotations.SerializedName
-
-data class SearchNoticeDataResponse(
-    @SerializedName("noticeList") val noticeList: List<SearchNoticeResponse>
-)
