--- app/src/main/java/com/ku_stacks/ku_ring/data/api/response/SubscribeListResponse.kt
@@ -1,15 +1,0 @@
-package com.ku_stacks.ku_ring.data.api.response
-
-import com.google.gson.annotations.SerializedName
-
-data class SubscribeListResponse(
-    @SerializedName(value = "message")
-    val resultMsg: String,
-    @SerializedName(value = "code")
-    val resultCode: Int,
-    @SerializedName(value = "data")
-    val categoryList: List<CategoryResponse>
-) {
-    val isSuccess: Boolean
-        get() = (resultCode == 200)
-}
