--- app/src/main/java/com/ku_stacks/ku_ring/repository/NoticeRepository.kt
@@ -1,21 +1,0 @@
-package com.ku_stacks.ku_ring.repository
-
-import androidx.paging.PagingData
-import com.ku_stacks.ku_ring.domain.Notice
-import io.reactivex.rxjava3.core.Completable
-import io.reactivex.rxjava3.core.Flowable
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.Flow
-
-interface NoticeRepository {
-    fun getNotices(type: String, scope: CoroutineScope): Flowable<PagingData<Notice>>
-    fun getSavedNotices(): Flow<List<Notice>>
-    fun getSavedNoticeList(): List<Notice>
-    fun insertNoticeAsOld(notice: Notice): Completable
-    fun updateNoticeToBeRead(articleId: String, category: String): Completable
-    suspend fun updateSavedStatus(articleId: String, category: String, isSaved: Boolean)
-    suspend fun updateNoticeToBeReadOnStorage(articleId: String, category: String)
-    suspend fun clearSavedNotices()
-    fun deleteAllNoticeRecord()
-    fun deleteSharedPreference()
-}
