+++ data/notice/src/main/java/com/ku_stacks/ku_ring/notice/api/response/CategoryResponse.kt
@@ -1,0 +1,9 @@
+package com.ku_stacks.ku_ring.notice.api.response
+
+import com.google.gson.annotations.SerializedName
+
+data class CategoryResponse(
+    @SerializedName("name") val name: String,
+    @SerializedName("hostPrefix") val shortName: String,
+    @SerializedName("korName") val koreanName: String,
+)
