+++ data/notice/src/main/java/com/ku_stacks/ku_ring/notice/api/response/DepartmentNoticeListResponse.kt
@@ -1,0 +1,7 @@
+package com.ku_stacks.ku_ring.notice.api.response
+
+data class DepartmentNoticeListResponse(
+    val code: Int,
+    val message: String,
+    val data: List<DepartmentNoticeResponse>,
+)
