+++ data/notice/src/main/java/com/ku_stacks/ku_ring/notice/api/response/SearchNoticeDataResponse.kt
@@ -1,0 +1,7 @@
+package com.ku_stacks.ku_ring.notice.api.response
+
+import com.google.gson.annotations.SerializedName
+
+data class SearchNoticeDataResponse(
+    @SerializedName("noticeList") val noticeList: List<SearchNoticeResponse>
+)
