+++ data/notice/src/main/java/com/ku_stacks/ku_ring/notice/api/response/SearchNoticeResponse.kt
@@ -1,0 +1,11 @@
+package com.ku_stacks.ku_ring.notice.api.response
+
+import com.google.gson.annotations.SerializedName
+
+data class SearchNoticeResponse(
+    @SerializedName("articleId") val articleId: String,
+    @SerializedName("postedDate") val postedDate: String,
+    @SerializedName("subject") val subject: String,
+    @SerializedName("baseUrl") val baseUrl: String,
+    @SerializedName("category") val category: String,
+)
